[refactor] CarResultView -> CarStatusDTO 이름 변경
CarRersultView 를 용도에 맞는 이름, 패키지로 이동
CarStatusDTO View의존 로직 제거
